+++ app/src/main/java/io/github/miuzarte/scrcpyforandroid/LockscreenPasswordActivity.kt
@@ -2,6 +2,7 @@ package io.github.miuzarte.scrcpyforandroid
 
 import android.content.Context
 import android.content.Intent
+import android.content.res.Configuration
 import android.os.Bundle
 import android.view.WindowManager
 import androidx.activity.compose.setContent
@@ -57,8 +58,22 @@ import top.yukonga.miuix.kmp.preference.SwitchPreference
 import top.yukonga.miuix.kmp.theme.MiuixTheme
 import top.yukonga.miuix.kmp.theme.MiuixTheme.colorScheme
 import top.yukonga.miuix.kmp.theme.MiuixTheme.textStyles
+import java.util.Locale
 
 class LockscreenPasswordActivity: FragmentActivity() {
+    override fun attachBaseContext(newBase: Context) {
+        val languageTag = MainActivity.getAppLanguageTag(newBase)
+        val wrappedContext =
+            if (languageTag.isNotEmpty()) {
+                val config = Configuration(newBase.resources.configuration)
+                config.setLocale(Locale.forLanguageTag(languageTag))
+                newBase.createConfigurationContext(config)
+            } else {
+                newBase
+            }
+        super.attachBaseContext(wrappedContext)
+    }
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         window.addFlags(WindowManager.LayoutParams.FLAG_SECURE)